+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 ### Added
 - Add minimal tests
 - Add multi parameters on Booking validator messages
+- Add static property access twig function
 
 ### Fixed
 - Fix js escaping on result